+++ components/button/stories/arg-types.js
@@ -23,6 +23,12 @@ export const argTypes = {
 			options: { 'true': '', 'false': undefined }
 		}
 	},
+	enlarged: {
+		control: {
+			type: 'inline-radio',
+			options: { 'true': '', 'false': undefined }
+		}
+	},
 	disabled: {
 		control: {
 			type: 'inline-radio',