+++ components/carousel/src/vwc-carousel.ts
@@ -35,6 +35,8 @@ export class VWCCarousel extends LitElement {
 	autoplay = true;
 	@query('.swiper-container')
 	private swiperContainer?: HTMLElement;
+	@query('.swiper-wrapper')
+	private swiperWrapper?: HTMLElement;
 	@query('.swiper-button-next')
 	private swiperButtonNext?: HTMLElement;
 	@query('.swiper-button-prev')
@@ -132,28 +134,30 @@ export class VWCCarousel extends LitElement {
 	}
 
 	private ensureStyleApplied() {
-		VWCCarousel.styles.forEach((style, index) => {
+		VWCCarousel.styles.forEach((styleResult, index) => {
 			const tmpId = `${CAROUSEL_STYLE_ID}-${index}`;
 			if (!document.head.querySelector(`#${tmpId}`)) {
 				const cs = document.createElement('style');
 				cs.id = tmpId;
 				cs.type = 'text/css';
-				cs.innerHTML = style.cssText;
+				cs.innerHTML = styleResult.cssText;
 				document.head.appendChild(cs);
 			}
 		});
 	}
 
 	private collectSlideRefs(swiper: Swiper): void {
-		for (let i = 0, l = swiper.slides.length; i < l; i++) {
-			this.slideRefs[i] = swiper.slides[i];
+		const slides = swiper.slides;
+		for (let i = 0, l = slides.length; i < l; i++) {
+			this.slideRefs[i] = slides[i];
 		}
 	}
 
 	private moveFirstIfNeeded(swiper?: Swiper): void {
 		const s = swiper ?? ((this as unknown) as Swiper);
-		if (s.slides.length > 2 && s.isEnd) {
-			const first = s.slides[0];
+		const slides = s.slides;
+		if (slides.length > 2 && s.isEnd) {
+			const first = slides[0];
 			s.removeSlide(0);
 			s.appendSlide(first);
 		}
@@ -161,9 +165,10 @@ export class VWCCarousel extends LitElement {
 
 	private moveLastIfNeeded(swiper?: Swiper): void {
 		const s = swiper ?? ((this as unknown) as Swiper);
-		if (s.slides.length > 2 && s.isBeginning) {
-			const last = s.slides[s.slides.length - 1];
-			s.removeSlide(s.slides.length - 1);
+		const slides = s.slides;
+		if (slides.length > 2 && s.isBeginning) {
+			const last = slides[slides.length - 1];
+			s.removeSlide(slides.length - 1);
 			s.prependSlide(last);
 		}
 	}
@@ -197,11 +202,11 @@ export class VWCCarousel extends LitElement {
 	}
 
 	private goToSlide(event: Event): void {
-		if (this.swiper && this.swiperPagination) {
+		if (this.swiper && this.swiperPagination && this.swiperWrapper) {
 			const logicalIndex = Array.from(this.swiperPagination.children).indexOf(
 				event.target as HTMLElement
 			);
-			const domIndex = Array.from(this.swiper.wrapperEl.children).indexOf(
+			const domIndex = Array.from(this.swiperWrapper.children).indexOf(
 				this.slideRefs[logicalIndex]
 			);
 			if (domIndex >= 0) {
@@ -212,6 +217,7 @@ export class VWCCarousel extends LitElement {
 
 	private calculateActiveIndex(swiper: Swiper): number {
 		const nai = swiper.activeIndex;
-		return this.slideRefs.indexOf(swiper.slides[nai]);
+		const slides = swiper.slides;
+		return this.slideRefs.indexOf(slides[nai]);
 	}
 }