+++ common/design-tokens/CHANGELOG.md
@@ -1,3 +1,33 @@
+# [0.15.0](https://github.com/vonage/vivid/compare/v0.14.0...v0.15.0) (2020-10-27)
+
+
+
+# [0.14.0](https://github.com/vonage/vivid/compare/v0.13.0...v0.14.0) (2020-10-25)
+
+
+### Bug Fixes
+
+* **scheme:** alternate variables support ([1c9c781](https://github.com/vonage/vivid/commit/1c9c781faaf71999040ea36fbeab969143ac40a2))
+* **scheme:** dark scheme correct variables ref ([fbaff84](https://github.com/vonage/vivid/commit/fbaff8409bdaca63306cf0740c786107c9e4175e))
+
+
+
+# [0.13.0](https://github.com/vonage/vivid/compare/v0.12.5...v0.13.0) (2020-10-12)
+
+
+
+## [0.12.5](https://github.com/vonage/vivid/compare/v0.12.4...v0.12.5) (2020-10-06)
+
+
+
+## [0.12.4](https://github.com/vonage/vivid/compare/v0.12.3...v0.12.4) (2020-10-06)
+
+
+
+## [0.12.3](https://github.com/vonage/vivid/compare/v0.12.2...v0.12.3) (2020-10-05)
+
+
+
 ## [0.12.2](https://github.com/vonage/vivid/compare/v0.12.1...v0.12.2) (2020-10-04)
 
 
@@ -6,7 +36,11 @@
 
 
 
-# [0.11.0](https://github.com/vonage/vivid/compare/v0.10.0...v0.11.0) (2020-09-29)
+# [0.11.0](https://github.com/vonage/vivid/compare/v0.11.2...v0.11.0) (2020-09-29)
+
+
+
+## [0.11.2](https://github.com/vonage/vivid/compare/v0.10.0...v0.11.2) (2020-09-29)
 
 
 ### Bug Fixes